switched to a jest pipeline
