+++ samples/azure-devops-ci-cd-spfx/azure-pipelines.yml
@@ -2,7 +2,7 @@ resources:
 - repo: self
 queue:
   name: Hosted VS2017
-  demands: 
+  demands:
   - npm
   - node.js
 
@@ -16,7 +16,6 @@ steps:
   displayName: 'Use Node 8.x'
   inputs:
     versionSpec: 8.x
-
     checkLatest: true
 
 
@@ -25,41 +24,38 @@ steps:
   displayName: 'npm install'
   inputs:
     workingDir: '$(workingDir)'
-
     verbose: false
 
 #start unit tests
-- task: Gulp@0
-  displayName: 'gulp test'
+- task: Npm@1
+  displayName: 'npm test'
   inputs:
-    gulpFile: '$(workingDir)/gulpfile.js'
+    command: custom
+    customCommand: 'test'
+    workingDir: '$(Build.SourcesDirectory)'
+    verbose: false
 
-    targets: test
+# Publish Test Results to Azure Pipelines/TFS
+- task: PublishTestResults@2
+  inputs:
+    testResultsFiles: 'temp/test/junit/junit.xml'
+    searchFolder: '$(Build.SourcesDirectory)'
 
-    publishJUnitResults: true
-
-    testResultsFiles: '**/test-*.xml'
-
-#publish test results
+# publish coverage test results
 - task: PublishCodeCoverageResults@1
-  displayName: 'Publish Code Coverage Results $(Build.SourcesDirectory)/$(workingDir)/temp/coverage/cobertura/cobertura.xml'
+  displayName: 'Publish Code Coverage Results $(Build.SourcesDirectory)/temp/test/cobertura-coverage.xml'
   inputs:
     codeCoverageTool: Cobertura
-
-    summaryFileLocation: '$(Build.SourcesDirectory)/$(workingDir)/temp/coverage/cobertura/cobertura.xml'
-
-    reportDirectory: '$(Build.SourcesDirectory)/$(workingDir)/temp/coverage/cobertura'
+    summaryFileLocation: '$(Build.SourcesDirectory)/temp/test/cobertura-coverage.xml'
+    reportDirectory: '$(Build.SourcesDirectory)/temp/test/'
 
 #bundle code with gulp
 - task: Gulp@0
   displayName: 'gulp bundle'
   inputs:
     gulpFile: '$(workingDir)/gulpfile.js'
-
     targets: bundle
-
     arguments: '--ship'
-
   continueOnError: true
 
 #package solution with gulp
@@ -67,9 +63,7 @@ steps:
   displayName: 'gulp package-solution'
   inputs:
     gulpFile: '$(workingDir)/gulpfile.js'
-
     targets: 'package-solution'
-
     arguments: '--ship'
 
 #copy files to artifact repository
@@ -77,7 +71,6 @@ steps:
   displayName: 'Copy Files to: $(build.artifactstagingdirectory)/drop'
   inputs:
     Contents: '**\*.sppkg'
-
     TargetFolder: '$(build.artifactstagingdirectory)/drop'
 
 #publish artifacts
@@ -85,6 +78,3 @@ steps:
   displayName: 'Publish Artifact:  drop'
   inputs:
     PathtoPublish: '$(build.artifactstagingdirectory)/drop'
-
-
-
