+++ samples/azure-devops-ci-cd-spfx/config/jest.config.json
@@ -1,0 +1,15 @@
+{
+  "preset": "@voitanos/jest-preset-spfx-react16",
+  "rootDir": "../src",
+  "coverageReporters": [
+    "text",
+    "json",
+    "lcov",
+    "text-summary",
+    "cobertura"
+  ],
+  "reporters": [
+    "default",
+    "jest-junit"
+  ]
+}
