--- samples/azure-devops-ci-cd-spfx/config/karma.config.js
@@ -1,30 +1,0 @@
-"use strict";
-var existingKarmaConfig = require('@microsoft/sp-build-web/lib/karma/karma.config');
-var junitReporter = require('karma-junit-reporter');
-
-module.exports = function (config) {
-  existingKarmaConfig(config);
-  config.reporters.push('junit');
-
-  config.set({
-    basePath: './..',
-  });
-
-  config.junitReporter = {
-    outputDir: 'temp/', // results will be saved as $outputDir/$browserName.xml
-    outputFile: 'test-results.xml', // if included, results will be saved as $outputDir/$browserName/$outputFile
-    suite: 'karma', // suite will become the package name attribute in xml testsuite element
-    useBrowserName: true, // add browser name to report and classes names
-  };
-  var coberturaSubDir = 'cobertura';
-  var coverageSubDir = 'lcov';
-  var coberturaFileName = 'cobertura.xml';
-  config.coverageReporter.reporters.push({type: 'cobertura', subdir: './' + coberturaSubDir, file: coberturaFileName});
-  config.coverageReporter.reporters.push({
-    type: 'lcov',
-    subdir: './' + coverageSubDir + '/',
-    file: 'lcov.info'
-  });
-  config.browserNoActivityTimeout = 60000;
-  config.plugins.push(junitReporter);
-};
