+++ samples/azure-devops-ci-cd-spfx/package.json
@@ -26,16 +26,22 @@
     "@microsoft/sp-tslint-rules": "1.8.0",
     "@microsoft/sp-webpart-workbench": "1.8.0",
     "@types/chai": "3.4.34",
-    "@types/chai-as-promised": "0.0.31",
     "@types/mocha": "2.2.38",
+    "@voitanos/jest-preset-spfx-react16": "^1.1.0",
     "ajv": "~5.2.2",
-    "chai-as-promised": "^6.0.0",
+    "chai": "^4.2.0",
     "gulp": "~3.9.1",
-    "karma-junit-reporter": "^1.2.0"
+    "jest": "^24.5.0",
+    "jest-junit": "^6.3.0"
   },
   "scripts": {
     "build": "gulp bundle",
     "clean": "gulp clean",
-    "test": "gulp test"
+    "test": "./node_modules/.bin/jest --config ./config/jest.config.json",
+    "test:watch": "./node_modules/.bin/jest --config ./config/jest.config.json --watchAll"
+  },
+  "jest-junit": {
+    "output": "temp/test/junit/junit.xml",
+    "usePathForSuiteName": "true"
   }
 }
